health: Add stubs for non-Linux platforms
