+++ pkg/health/server/health.go
@@ -15,32 +15,14 @@
 package server
 
 import (
-	"strconv"
-
 	healthModels "github.com/cilium/cilium/api/v1/health/models"
 	. "github.com/cilium/cilium/api/v1/health/server/restapi"
 	ciliumModels "github.com/cilium/cilium/api/v1/models"
 	"github.com/cilium/cilium/pkg/api"
 	"github.com/cilium/cilium/pkg/client"
-	"golang.org/x/sys/unix"
 
 	"github.com/go-openapi/runtime/middleware"
 )
-
-func dumpLoad() (*healthModels.LoadResponse, error) {
-	var info unix.Sysinfo_t
-	err := unix.Sysinfo(&info)
-	if err != nil {
-		return nil, err
-	}
-
-	scale := float64(1 << unix.SI_LOAD_SHIFT)
-	return &healthModels.LoadResponse{
-		Last1min:  strconv.FormatFloat(float64(info.Loads[0])/scale, 'f', 2, 64),
-		Last5min:  strconv.FormatFloat(float64(info.Loads[1])/scale, 'f', 2, 64),
-		Last15min: strconv.FormatFloat(float64(info.Loads[2])/scale, 'f', 2, 64),
-	}, nil
-}
 
 type getHealthz struct {
 	*Server
