+++ pkg/health/server/health_unspecified.go
@@ -1,0 +1,27 @@
+// Copyright 2017-2021 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build !linux
+
+package server
+
+import (
+	"errors"
+
+	healthModels "github.com/cilium/cilium/api/v1/health/models"
+)
+
+func dumpLoad() (*healthModels.LoadResponse, error) {
+	return nil, errors.New("not implemented")
+}
